+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 29.07.2023
+
+[Diff](https://github.com/elk-language/go-prompt/compare/v1.1.3...elk-language:go-prompt:v1.1.4)
+
+### Added
+- `func (*prompt.Document).PreviousLineIndentSpaces() int`
+- `func (*prompt.Document).PreviousLineIndentLevel(indentSize int) int`
+- `func (*prompt.Document).PreviousLine() (s string, ok bool)`
+
 ## [1.1.3] - 29.07.2023
 
 [Diff](https://github.com/elk-language/go-prompt/compare/v1.1.2...elk-language:go-prompt:v1.1.3)